// Server/src/api/li/interfaces/Types/IItem.ts
export interface IItem {
  altura: number
  disponibilidade: number
  id: number
  largura: number
  linha: number
  nome: string
  pedido: string
  peso: string
  preco_cheio: string
  preco_custo: string
  preco_promocional: string
  preco_subtotal: string
  preco_venda: string
  produto: string
  produto_pai: string
  profundidade: number
  quantidade: string
  sku: string
  tipo: string
}